requirements.txt
Django==3.0.5
django-allauth==0.41.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
